Renamed tests.
