+++ src/persistence/secondary/secondary-entity-manager.ts
@@ -1,7 +1,7 @@
 import { Entity } from '@antjs/ant-js';
-import { SecondaryEntityManager } from '@antjs/ant-js/src/persistence/secondary/secondary-entity-manager';
+import { SecondaryEntityManager as AntJsSecondaryEntityManager } from '@antjs/ant-js/src/persistence/secondary/secondary-entity-manager';
 
-export interface ISqlSecondaryEntityManager<TEntity extends Entity> extends SecondaryEntityManager<TEntity> {
+export interface SecondaryEntityManager<TEntity extends Entity> extends AntJsSecondaryEntityManager<TEntity> {
   /**
    * Deletes an entity from its id.
    * @param id Id of the entity to delete.
