+++ src/test/AllTests.ts
@@ -1,16 +1,16 @@
 import { Test } from '@antjs/ant-js/src/testapi/api/test';
 import { AntTest } from './AntTest';
-import { AntSqlManagerTest } from './api/AntSqlManagerTest';
-import { AntSqlModelManagerTest } from './api/AntSqlModelManagerTest';
-import { QueryConfigFactoryTest } from './api/QueryConfigFactoryTest';
+import { AntSqlManagerTest } from './api/ant-sql-manager-test';
+import { AntSqlModelManagerTest } from './api/ant-sql-model-manager-test';
+import { QueryConfigFactoryTest } from './api/query-config-factory-test';
 import { dbServerAwaiterManager } from './await/DbServerAwaiterManager';
 import { AntSqlModelTest } from './model/AntSqlModelTest';
 import { RedisWrapper } from './persistence/primary/RedisWrapper';
-import { SqlModelManagerTest } from './persistence/primary/SqlModelManagerTest';
-import { AntSqlSecondaryEntityManagerTest } from './persistence/secondary/AntSqlSecondaryEntityManagerTest';
+import { AntSqlPrimaryModelManagerTest } from './persistence/primary/ant-sql-primary-model-manager-test';
+import { SqlSecondaryEntityManagerTest } from './persistence/secondary/sql-secondary-entity-manager-test';
 import { DBConnectionWrapper } from './persistence/secondary/DBConnectionWrapper';
 import { DBTestManager } from './persistence/secondary/DBTestManager';
-import { SecondaryEntityManagerHelperTest } from './persistence/secondary/SecondaryEntityManagerHelperTest';
+import { SecondaryEntityManagerHelperTest } from './persistence/secondary/secondary-entity-manager-helper-test';
 
 export class AllTest implements Test {
   /**
@@ -36,7 +36,7 @@ export class AllTest implements Test {
     const dBConnectionWrapper = new DBConnectionWrapper();
 
     const fakeConnection = dBConnectionWrapper.fakeConnection;
-    new SqlModelManagerTest(fakeConnection, 'fake').performTests();
+    new AntSqlPrimaryModelManagerTest(fakeConnection, 'fake').performTests();
 
     for (const config of dBConnectionWrapper.config) {
       const connection = config.connection;
@@ -46,7 +46,7 @@ export class AllTest implements Test {
 
       new AntSqlModelManagerTest(connection, connection.client.driverName).performTests();
 
-      new AntSqlSecondaryEntityManagerTest(
+      new SqlSecondaryEntityManagerTest(
         deleteAllTablesPromise,
         connection,
         connection.client.driverName,
